+++ packages/framework/src/Markdown/Processing/BladeDownProcessor.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Markdown;
+namespace Hyde\Markdown\Processing;
 
 use Hyde\Markdown\Contracts\MarkdownPostProcessorContract;
 use Hyde\Markdown\Contracts\MarkdownPreProcessorContract;
