+++ packages/framework/tests/Feature/Services/Markdown/CodeblockFilepathProcessorTest.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Feature\Services\Markdown;
 
-use Hyde\Markdown\CodeblockFilepathProcessor;
+use Hyde\Markdown\Processing\CodeblockFilepathProcessor;
 use Hyde\Testing\TestCase;
 
 /**
- * @covers \Hyde\Markdown\CodeblockFilepathProcessor
+ * @covers \Hyde\Markdown\Processing\CodeblockFilepathProcessor
  */
 class CodeblockFilepathProcessorTest extends TestCase
 {
@@ -37,7 +37,7 @@ class CodeblockFilepathProcessorTest extends TestCase
             $markdown = "\n```php\n{$pattern}foo.php\necho 'Hello World';\n```";
             $expected = "\n<!-- HYDE[Filepath]foo.php -->\n```php\necho 'Hello World';\n```";
 
-            $this->assertEquals($expected, \Hyde\Markdown\CodeblockFilepathProcessor::preprocess($markdown));
+            $this->assertEquals($expected, \Hyde\Markdown\Processing\CodeblockFilepathProcessor::preprocess($markdown));
         }
     }
 
@@ -55,10 +55,10 @@ class CodeblockFilepathProcessorTest extends TestCase
             $markdown = "\n```{$language}\n// filepath: foo.{$language}\nfoo\n```";
             $expected = "\n<!-- HYDE[Filepath]foo.{$language} -->\n```{$language}\nfoo\n```";
 
-            $this->assertEquals($expected, \Hyde\Markdown\CodeblockFilepathProcessor::preprocess($markdown));
+            $this->assertEquals($expected, \Hyde\Markdown\Processing\CodeblockFilepathProcessor::preprocess($markdown));
         }
 
-        $this->assertEquals($expected, \Hyde\Markdown\CodeblockFilepathProcessor::preprocess($markdown));
+        $this->assertEquals($expected, \Hyde\Markdown\Processing\CodeblockFilepathProcessor::preprocess($markdown));
     }
 
     public function test_preprocess_accepts_multiple_input_blocks()
@@ -89,7 +89,7 @@ class CodeblockFilepathProcessorTest extends TestCase
         ```
         MD;
 
-        $this->assertEqualsIgnoringLineReturnType($expected, CodeblockFilepathProcessor::preprocess($markdown));
+        $this->assertEqualsIgnoringLineReturnType($expected, \Hyde\Markdown\Processing\CodeblockFilepathProcessor::preprocess($markdown));
     }
 
     public function test_preprocess_accepts_multi_line_codeblocks()
@@ -114,7 +114,7 @@ class CodeblockFilepathProcessorTest extends TestCase
         ```
         MD;
 
-        $this->assertEqualsIgnoringLineReturnType($expected, \Hyde\Markdown\CodeblockFilepathProcessor::preprocess($markdown));
+        $this->assertEqualsIgnoringLineReturnType($expected, \Hyde\Markdown\Processing\CodeblockFilepathProcessor::preprocess($markdown));
     }
 
     public function test_space_after_filepath_is_optional()
@@ -136,8 +136,8 @@ class CodeblockFilepathProcessorTest extends TestCase
         ```
         MD;
 
-        $this->assertEqualsIgnoringLineReturnType(\Hyde\Markdown\CodeblockFilepathProcessor::preprocess($expected),
-            \Hyde\Markdown\CodeblockFilepathProcessor::preprocess($markdown));
+        $this->assertEqualsIgnoringLineReturnType(\Hyde\Markdown\Processing\CodeblockFilepathProcessor::preprocess($expected),
+            \Hyde\Markdown\Processing\CodeblockFilepathProcessor::preprocess($markdown));
     }
 
     public function test_processor_expands_filepath_directive_in_standard_codeblock()
@@ -151,7 +151,7 @@ class CodeblockFilepathProcessorTest extends TestCase
         <pre><code class="language-html"><small class="filepath"><span class="sr-only">Filepath: </span>foo.html</small></code></pre>
         HTML;
 
-        $this->assertEqualsIgnoringLineReturnType($expected, CodeblockFilepathProcessor::postprocess($html));
+        $this->assertEqualsIgnoringLineReturnType($expected, \Hyde\Markdown\Processing\CodeblockFilepathProcessor::postprocess($html));
     }
 
     public function test_processor_expands_filepath_directive_in_torchlight_codeblock()
@@ -165,7 +165,7 @@ class CodeblockFilepathProcessorTest extends TestCase
         <pre><code class="torchlight"><!-- Syntax highlighted by torchlight.dev --><small class="filepath"><span class="sr-only">Filepath: </span>foo.html</small><div class="line"><span class="line-number">1</span>&nbsp;</div></code></pre>
         HTML;
 
-        $this->assertEqualsIgnoringLineReturnType($expected, CodeblockFilepathProcessor::postprocess($html));
+        $this->assertEqualsIgnoringLineReturnType($expected, \Hyde\Markdown\Processing\CodeblockFilepathProcessor::postprocess($html));
     }
 
     protected function assertEqualsIgnoringLineReturnType(string $expected, string $actual)
