+++ packages/framework/tests/Feature/Services/Markdown/ShortcodeProcessorTest.php
@@ -5,17 +5,16 @@ declare(strict_types=1);
 namespace Hyde\Framework\Testing\Feature\Services\Markdown;
 
 use Hyde\Markdown\Contracts\MarkdownShortcodeContract;
-use Hyde\Markdown\ShortcodeProcessor;
 use Hyde\Testing\TestCase;
 
 /**
- * @covers \Hyde\Markdown\ShortcodeProcessor
+ * @covers \Hyde\Markdown\Processing\ShortcodeProcessor
  */
 class ShortcodeProcessorTest extends TestCase
 {
     public function test_constructor_discovers_default_shortcodes()
     {
-        $shortcodes = (new ShortcodeProcessor('foo'))->shortcodes;
+        $shortcodes = (new \Hyde\Markdown\Processing\ShortcodeProcessor('foo'))->shortcodes;
 
         $this->assertCount(4, $shortcodes);
         $this->assertContainsOnlyInstancesOf(MarkdownShortcodeContract::class, $shortcodes);
@@ -23,7 +22,7 @@ class ShortcodeProcessorTest extends TestCase
 
     public function test_discovered_shortcodes_are_used_to_process_input()
     {
-        $processor = new \Hyde\Markdown\ShortcodeProcessor('>info foo');
+        $processor = new \Hyde\Markdown\Processing\ShortcodeProcessor('>info foo');
 
         $this->assertEquals('<blockquote class="info">foo</blockquote>',
             $processor->run());
@@ -31,7 +30,7 @@ class ShortcodeProcessorTest extends TestCase
 
     public function test_string_without_shortcode_is_not_modified()
     {
-        $processor = new \Hyde\Markdown\ShortcodeProcessor('foo');
+        $processor = new \Hyde\Markdown\Processing\ShortcodeProcessor('foo');
 
         $this->assertEquals('foo', $processor->run());
     }
@@ -39,12 +38,12 @@ class ShortcodeProcessorTest extends TestCase
     public function test_process_static_shorthand()
     {
         $this->assertEquals('<blockquote class="info">foo</blockquote>',
-            \Hyde\Markdown\ShortcodeProcessor::preprocess('>info foo'));
+            \Hyde\Markdown\Processing\ShortcodeProcessor::preprocess('>info foo'));
     }
 
     public function test_shortcodes_can_be_added_to_processor()
     {
-        $processor = new \Hyde\Markdown\ShortcodeProcessor('foo');
+        $processor = new \Hyde\Markdown\Processing\ShortcodeProcessor('foo');
 
         $processor->addShortcode(new class implements MarkdownShortcodeContract
         {
